chore: comprehensive project normalization and tooling setup
Docs:
- Rewrite AGENTS.md (535 → 170 lines): project-specific, OpenSpec+AI workflow, Chinese-first
- Rewrite CLAUDE.md: focused Claude guidance, no redundancy with AGENTS.md
- Simplify CONTRIBUTING.md (378 → ~50 lines): minimal, single-person project appropriate
- Update openspec/config.yaml: full workflow, branching strategy, AI tool integration

Tooling:
- Add .clangd: LSP config for C++/CUDA (clangd, works with VSCode/neovim/copilot)
- Add .githooks/pre-commit: auto clang-format enforcement
- Add .github/copilot-instructions.md: Chinese-first, project-specific Copilot behavior

CI:
- Merge configure-no-cuda + build-cpu into single build-verify job (removes redundancy)

OpenSpec:
- Add openspec/changes/active/project-completion/proposal.md: project closeout tasks

Bug fixes:
- fix(tests): raise pagerank/benchmark property test iterations from 50 → 100 (spec compliance)
- fix(bandwidth): explicit static_cast<size_t> for int*sizeof to prevent implicit narrowing

+++ src/bandwidth.cpp
@@ -44,12 +44,12 @@ BandwidthMetrics compute_bandwidth_csr(const CSRMatrix* A, float elapsed_ms) {
     // 写入: y (num_rows * sizeof(float))
 
     size_t bytes_read = 0;
-    bytes_read += A->nnz * sizeof(float);           // values
-    bytes_read += A->nnz * sizeof(int);             // col_indices
-    bytes_read += (A->num_rows + 1) * sizeof(int);  // row_ptrs
-    bytes_read += A->num_cols * sizeof(float);      // x vector
+    bytes_read += static_cast<size_t>(A->nnz) * sizeof(float);           // values
+    bytes_read += static_cast<size_t>(A->nnz) * sizeof(int);             // col_indices
+    bytes_read += static_cast<size_t>(A->num_rows + 1) * sizeof(int);    // row_ptrs
+    bytes_read += static_cast<size_t>(A->num_cols) * sizeof(float);      // x vector
 
-    size_t bytes_write = A->num_rows * sizeof(float);  // y vector
+    size_t bytes_write = static_cast<size_t>(A->num_rows) * sizeof(float);  // y vector
 
     size_t total_bytes = bytes_read + bytes_write;
 
@@ -78,11 +78,11 @@ BandwidthMetrics compute_bandwidth_ell(const ELLMatrix* A, float elapsed_ms) {
     size_t ell_size = static_cast<size_t>(A->num_rows) * A->max_nnz_per_row;
 
     size_t bytes_read = 0;
-    bytes_read += ell_size * sizeof(float);     // values
-    bytes_read += ell_size * sizeof(int);       // col_indices
-    bytes_read += A->num_cols * sizeof(float);  // x vector
+    bytes_read += ell_size * sizeof(float);                            // values
+    bytes_read += ell_size * sizeof(int);                              // col_indices
+    bytes_read += static_cast<size_t>(A->num_cols) * sizeof(float);   // x vector
 
-    size_t bytes_write = A->num_rows * sizeof(float);  // y vector
+    size_t bytes_write = static_cast<size_t>(A->num_rows) * sizeof(float);  // y vector
 
     size_t total_bytes = bytes_read + bytes_write;
